core/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Base de datos
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Cache / Queue
redis==5.2.0
celery==5.4.0

# Vector DB
qdrant-client==1.12.0

# ML / Embeddings — fastembed reemplaza torch+sentence-transformers
# torch pesa 2GB y rompe el build en Render free tier
fastembed>=0.4.2

# Seguridad
passlib[bcrypt]==1.7.4
cryptography==43.0.3

# HTTP
httpx==0.27.2
aiohttp==3.11.0

# Validación
pydantic[email]==2.10.0
pydantic-settings==2.6.0

# LLM providers
anthropic==0.39.0
openai==1.56.0

# Utilidades
python-dotenv==1.0.1
structlog==24.4.0
prometheus-client==0.21.0
docker==7.1.0

# Observabilidad
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-instrumentation-fastapi==0.49b0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# User added dependencies
PyJWT==2.10.0
PyGithub==2.5.0
boto3==1.35.0
